fix(SnackBar): fix some mistakes with types and annotations
closes #1011
