+++ src/components/SnackBar/__stories__/examples/SnackBarExampleTimer/SnackBarExampleTimer.tsx
@@ -15,7 +15,7 @@ import { SnackBarItemShowProgressProp, SnackBarItemStatus } from '../../../types
 const cnSnackBarExampleTimer = cn('SnackBarExampleTimer');
 
 type Item = {
-  key: string;
+  key: number;
   message: string;
   icon?: IconComponent;
   buttons?: string[];
@@ -58,7 +58,7 @@ export const SnackBarExampleTimer: React.FC = () => {
   ) => () => {
     const key = items.length + 1;
     const item: Item = {
-      key: key.toString(),
+      key,
       message: `Сейчас эта штука закроется ${key}`,
       status,
       icon: getItemIconByStatus(status),
